Add isInBetween to gameBoard
Figuring out how to check if the clicked cell is between two claimed
cells was a major stalling point. Use a loop in isInBetween to check
three elements at a time: the currently clicked (with the i and j
parameters), and two adjacent claimed elements ([k, l], [m, n]). There
are two deviations to track at a time. The deviations between row and
column with one adjacent and the deviations with another adjacent. If
the two deviations are added together and the result is [0, 0], the
clicked element is in between.