Add board clear after win and fix score bug
Add optional parameter to DisplayController.clearBoard() so that
the function can discern whether it is being called after a win. If so,
the board clear animation is played. If not, it is called when the
player presses the home button in which case no animation is needed.
Also, there was a bug where the scored from the previous match would
still be showing even when the player pressed the home button. Fix this
so that it is always 0-0 when a new game is begun.