Move query of gameBoard into displayController.win()
gameBoard is cloned every time the home button is pressed to avoid
multiple event listeners being added to the same cell. This causes
the need for gameBoard to be queried each time it needs to be used,
because if it is stored in a closure it will no longer be valid once
game.removeListeners() clones it. Move the query into .win() so that
the current version of the board is accessed every time.