Fix bot.move() glitch
Previously, a setTimeout() was set inside bot.move() in order to give
the program time to clear the board after a win. However, whenever there
was a match in which a player won with the board entirely full, an error
would be thrown saying that there were no available spaces. This was
because cells was being passed to bot.move() before the board was
cleared. Move setTimeout() to when bot.move() is CALLED instead of
inside it, making sure that the correct version of cells is always being
passed.